requirements.txt
requests==2.25.*
pandas==1.2.*
django==3.1.*
djangorestframework==3.12.*
django-filter==2.4.*
